+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # OSIM Changelog
+## [Unreleased]
+### Added
+* Add manual component suggestions for Source Component field (`OSIDB-4996`)
+
 ## [2026.5.1]
 ### Added
 * Add support for Business Unit (BU) label type for flaw triage (`OSIDB-4982`)